For platform downloads, support git cloning in addition to gitweb snapshots.
To use this, prefix the URL in platforms.js with a git protocol.
git:// is accepted as is, others go by the common convention of
having "git+" prefixed to the actual URL. So we end up with
git+ssh://, git+http[s]://, git+ftp[s]://, git+rsync:// and
git+file:// for git URLs that you want to clone.

(The common form git@host.xz:path/to/repo.git is also an SSH URL,
see git help clone for details.)