Remove npm-shrinkwrap.json
After some discussion, looks like the lib doesn't need a shrinkwrap. The one
in end-user tools (CLI, plugman, etc.) should take care of freezing the deps
for the tools to work.

In CLI, the shrinkwrap file be created temporarily for release and
then removed again, updating the coho docs with this:
http://github.com/apache/cordova-coho/blob/master/docs/tools-release-process.md