CB-6512: platform add <path> was using wrong www/cordova.js
cordova prepare nukes the www dir under platform on each prepare, so it was
using www/cordova.js from the cached platfrom files to restore it. For this to
work, prepare was starting with a lazy_load(platform) to make sure it has the
platform files cached.

For the case of platform add <path> this was downloading a stock version of
platform files instead of using the ones in <path>.

Some time ago see CB-5063 cordova started keeping a copy of cordova.js under
/platfroms/<platfrom>/platfrom_www/ but the cached files were still needed for
older projects that didn't yet have platform_www.

This commit removes the lazy_load before prepare and counts on platform_www to
exist. It also adds platform_www during `platform add`.