url and local save/restore support
Implements save and restore for plugins that are coming from local file system or from a git repo.
Also makes a change so that the dependency plugins are not saved during a save. Makes a change
on the fetch.js so that .fetch.json now accurately records the type for plugin installations from registry
as type 'registry'