Fix xcode project manipulation on Windows OS
When adding plugins on Windows, plugman will insert incorrect paths, which also contain backslashes into the xcode project files. This breaks builds of the iOS app in Xcode and requires manual editing of the project file.

On all non-Windows platforms, where the path separator is already `/`, the performed replace should be a noop, on Windows it fixes the paths in the project file.