CB-8053: Including a project reference in a plugin fails on Windows platform.
When working with the Windows platform, the existing logic to add a project reference was expecting two things:

1. There would only be one solution file.
2. That solution file would only contain a single project.

This is no longer true - there are now two solution files (one for VS 2012 and one for VS 2013). The VS 2013 solution is for a universal project, and contains five projects.

The change here is to make the code (in cordova-lib/src/util/windows/jsproj.js) which modifies the solution file more versatile by:

1. Handling any number of solution files (iterate all solution files we find).
2. Handling any number of projects in the solution (add the build depedency to every jsproj in the solution).

Includes matching changes to removeProjectReference(), and fixes a call in cordova-lib/src/plugman/platforms/windows.js to pass the correct path for plugin_dir.