CB-8651 Restoring platforms causes plugin install to be triggered twice.
If you add a platform and there are plugins listed in config.xml, those plugins get added twice. One during prepare(), when we add all plugins listed in config.xml, and once in platform.add(), when we add all installed plugins to the new platform.

Both these cases are needed in different scenarios, so the easiest fix seemed to be to check for and ignore plugins that have already been installed for a platform, in installPluginsForNewPlatform().