CB-8807 Platform Add fails to add plugins with variables.
When a plugin that uses variables is automatically installed for a new platform, it fails with the following exception:

    Installing "[plugin]" for [platform]
    Failed to install '[plugin]':Error: Variable(s) missing: [variables]
        at .../cordova-lib/src/plugman/install.js:299:23
        at ...

The problem here is that we are looking in config.xml for information about the plugins variables, but the plugin is only saved to config.xml if it was added with the --save flag. This fix updates the logic to look in fetch.json instead.