CB-9051 Plugins don't get re-added if platforms folder deleted.
To avoid attempting to add a plugin twice while adding a platform (a scenario that can occur now with automatic restore of plugins and platforms), when adding plugins to a newly added platform we check to see if the plugin has already been added by calling plugman's PlatformJson.isPluginInstalled(), which looks in plugins/<platform>.json to see if the plugin has been installed for a platform.

If you delete a platform's folder, it's plugins/<platform>.json file still exists, so if you add the platform again, we think all plugins are already installed for it and so don't try to install them.

Simple fix is to delete the platform's plugins/<platform>.json file before adding a platform, so we start with a clean slate.