CB-9108 Handle version ranges when add platform with --usegit.
Adding a platform using a version range didn't work with --usegit, since it would try to use that version range as a git tag. This was a problem if you added the platform using 'platform add <platform>@<range>', and also because our pinned platforms now use tilde versions (which are shorthand for a version range).

Fix is to use npm to find the most recent version that matches the range, and use that.

We need to do the same when installing from npm, when we do our check to see if the requested version has previously been downloaded using git.