Fix four failing tests on Windows.
Two tests fail in a non-admin command window because they rely on creating symlinks. Since creating the symlinks isn't the purpose of the test but just setup for the test, if that fails and the platform is win32, we now just bail on the test.

Another test failed because a folder unexpectedly exists when run from a non-admin prompt. I'm not 100% sure why the folder exists, but the test was actually bogus... the test was actually ends up testing that we throw if src does not exist (something that we verify in an earlier test), while the intent was to verify we throw if dest *already* exists. It is copyNewFile() that throws in that scenario, so I updated the test to check for that.

Finally, a fourth test failed on Windows (admin or non-admin) because of different line endings. Changed that to handle any line ending.