CB-9147 Adding a platform via caret version adds latest rather than the latest matching.
Internally we are using a version of npm that doesn't understand caret versions, so it just gets the latest. Missed this previously because when I tried this, the latest version *was* the latest matching version.

The same problem also applies to plugins.

Slightly different fix in each case:

For platforms, we already had the latest matching version as a result of our call to getLatestMatchingNpmVersion(), so we just use that.

For plugins, in this specific scenario (a caret version) we turn it into a standard range using a version of semver that understands caret ranges (so '^1.2.3' becomes '>=1.2.3 <2.0.0') and use that instead.

Also, I noticed that we failed to find the version in config.xml if you added it using the new form (cordova-plugin-camera, for example), but it was stored in config.xml using the id, so fixed that.