CB-9813 Keep module-to-plugin mapping at hand.
This information is handy for various use-cases. E.g. for preview application to
provide only the plugins which are used by the application (instead of all of
them, since they are built-in in the preview app on the device). The other is
simulation (taco-simulation does need this info).

Instead of trying to (unreliably) recover this information from module list,
just keep track of it when it's known (at the prepare stage).