Downgrade error to warning
Replace error with warning message if search paths differ.
(which does happen with `cca` tool from the MobileChromeApps project)

Generally, if the `localPlugins` cache variable has been previously set,
(via programmatically calling plugin add), then subsequent calls with
differing paths will cause the process to exit, which seems a bit extreme.
As the comments suggest, this is simply not implemented, not a failure
state, and can be safely passed.

 This closes #397