Break cycle between cordova/util and platforms/platforms
We do this by moving `cordova/util.hostSupports` to the `platforms`
sub-package. To not further pollute `platforms/platforms`, we introduce
a new facade module for the the `platforms` sub-package.

`platforms/platforms` remains for the time being, especially because it
is part of our public interface (see `cordova-lib.js`).

This commit also simplifies the platforms.hostSupports utility function.
The support for indicating no host requirements by setting `hostos: '*'`
is dropped in favor of letting `hostos` undefined.