Prevent masking of errors during testing (#771)
Node.js throws an Error if the error event is emitted on a EventEmitter instance that has no handlers attached for it. This is usually the case during testing and often masks the actual error that is causing the issue.

So we attach a listener to just log error events to console. During passing tests runs, no additional output is generated.