Have plugman.createPackageJson create file in plugin dir, not in cwd (#773)
Right now `plugman createpackagejson <plugin dir>` reads values from `plugin.xml` in `<plugin dir>` but creates the `package.json` in `process.cwd()`. This seems to be undocumented behavior, and is considered a bug.

With this patch, we always create the `package.json` in `<plugin dir>` next to `plugin.xml` instead.